refactor create asset page
